backend/requirements.txt
# FlowCraft AI - Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.2.0

# OCR & Document Processing
easyocr==1.7.0
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
pdf2image==1.16.3
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.1.0

# AI & Machine Learning
requests==2.31.0
numpy==1.26.4
pandas==2.1.4

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# System Monitoring
psutil==5.9.6

# Utilities
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production
gunicorn==21.2.0
